stage_2/requirements_stage2.txt
pandas>=2.1.0
pyyaml>=6.0.1
pyarrow>=15.0.0
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.1
